backend/requirements.txt
fastapi
uvicorn
pymongo
motor  # Async MongoDB driver
python-dotenv
pydantic[email]
passlib[bcrypt]
datetime
PyJWT 